requirements.txt

fastapi
uvicorn
python-dotenv
openai
requests
python-telegram-bot
